TILEAPI/requirements_v3.txt
# Requirements for v3.py - Hugging Face API (Lightweight!)
# No PyTorch needed - works with any system!

Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
gradio-client>=0.10.0  # Official HF client
